+++ .env.example
@@ -92,12 +92,9 @@ FIREBASE_SESSION_LIFETIME=432000
 
 # Firebase Emulator Settings (for development)
 FIREBASE_USE_EMULATOR=false
-FIREBASE_AUTH_EMULATOR_HOST=localhost
-FIREBASE_AUTH_EMULATOR_PORT=9099
-FIREBASE_DATABASE_EMULATOR_HOST=localhost
-FIREBASE_DATABASE_EMULATOR_PORT=9000
-FIREBASE_FIRESTORE_EMULATOR_HOST=localhost
-FIREBASE_FIRESTORE_EMULATOR_PORT=8080
+FIREBASE_AUTH_EMULATOR_HOST=127.0.0.1:9099
+FIREBASE_DATABASE_EMULATOR_HOST=127.0.0.1:9000
+FIREBASE_FIRESTORE_EMULATOR_HOST=127.0.0.1:8080
 
 # Multi-tenancy Configuration
 MULTI_TENANT_ENABLED=true
